server/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework_simplejwt==5.5.1
channels
daphne
numpy
tensorflow
pandas
scikit-learn
torch